llm-backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.3
attrs==25.3.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.7.0
exceptiongroup==1.2.2
Flask==3.1.0
flask-cors==5.0.1
fonttools==4.56.0
frozenlist==1.5.0
groq==0.22.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
langchain==0.3.22
langchain-community==0.3.20
langchain-core==0.3.50
langchain-experimental==0.3.4
langchain-groq==0.3.2
langchain-text-splitters==0.3.7
langsmith==0.3.23
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
minio==7.2.15
multidict==6.2.0
mypy-extensions==1.0.0
numpy==2.2.4
orjson==3.10.16
packaging==24.2
pandas==2.2.3
patsy==1.0.1
pillow==11.1.0
propcache==0.3.1
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.1
pydantic-settings==2.8.1
pydantic_core==2.33.0
pymongo==4.11.3
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
scipy==1.15.2
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
statsmodels==0.14.4
tenacity==9.1.2
typing==3.7.4.3
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
Werkzeug==3.1.3
yarl==1.18.3
zstandard==0.23.0
